backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
node==2.0.0
numpy==2.4.4
odict==2.0.0
openpyxl==3.1.5
pandas==3.0.2
plumber==2.0.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
zope.component==7.1
zope.deferredimport==6.1.1
zope.deprecation==6.0
zope.event==6.1
zope.hookable==8.2
zope.interface==8.2
zope.lifecycleevent==6.0
zope.proxy==7.1